+++ tests/_mocks/MockRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Swis\JsonApi\Client\Tests\Mocks;
+
+use Swis\JsonApi\Client\Repository;
+
+class MockRepository extends Repository
+{
+    protected $endpoint = 'mocks';
+}