+++ CHANGELOG.md
@@ -9,12 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Add headers to request methods in `DocumentClientInterface`.
-N.B. This is a breaking change if you implement the interface yourself or extend the `DocumentClient`.
+N.B. This is a breaking change if you implement the interface yourself or extend the `DocumentClient`. [#34](https://github.com/swisnl/json-api-client/pull/34)
 
 ### Removed
 
 * Removed obsolete `ItemDocumentSerializer` in favor of `JsonSerializable`.
 N.B. This is a breaking change if you use this class directly, construct the `DocumentClient` yourself or have overwritten `\Swis\JsonApi\Client\Providers\ServiceProvider::registerClients`. The `ItemDocument` can now be serialized using its `jsonSerialize` method.
+
+### Fixed
+
+* Do not fail on, but skip relationships without data [#38](https://github.com/swisnl/json-api-client/pull/38)
 
 ## [0.11.0] - 2018-12-21